% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-prop.R
\name{get_prop}
\alias{get_prop}
\title{Simple Binomial Proportion}
\usage{
get_prop(a, n, z = 0.05)
}
\arguments{
\item{a}{Numerator}

\item{n}{Denominator}

\item{z}{Desired confidence interval/limits. Defaults to 0.05 for 95\%
confidence interval/limits}
}
\value{
An object of class \code{binom} providing point estimate of simple
    binomial proportion and estimate of various confidence interval/limits
    calculation approaches
}
\description{
This function provides confidence limits for simple (binomial) proportions by
several different methods.
}
\details{
The numerator must be smaller than the denominator and both must be positive numbers.
}
\examples{
get_prop(a = 10, n = 11)

}
