% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-smr.R
\name{get_smr}
\alias{get_smr}
\title{The Standardized Mortality Ratio (SMR) is the ratio of observed to the
expected number of deaths in the study population under the assumption that
the mortality rates for the study population are the same as those for the
general population. This fumction estimates SMR and various confidence
intervals and tests for statistical significance.}
\usage{
get_smr(a, b, ci = 0.05)
}
\arguments{
\item{a}{The observed number of deaths}

\item{b}{The expected number of deaths}

\item{z}{Desired confidence interval/limits. Defaults to 0.05 for 95\%
confidence interval/limits}
}
\value{
An object of class \code{smr} providing point estimate of SMR,
    estimates of various confidence interval/limits calculation approaches,
    and results of tests for statistical significance
}
\description{
The Standardized Mortality Ratio (SMR) is the ratio of observed to the
expected number of deaths in the study population under the assumption that
the mortality rates for the study population are the same as those for the
general population. This fumction estimates SMR and various confidence
intervals and tests for statistical significance.
}
\examples{
get_smr(a = 4, b = 3.3)

}
