% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rOpenEpi.R
\docType{package}
\name{rOpenEpi}
\alias{rOpenEpi}
\alias{rOpenEpi-package}
\title{R Port of Open Source Epidemiologic Statistics for Public Health}
\description{
This package is the R port for the free and open source software OpenEpi
<https://www.openepi.com>. OpenEpi provides statistics for counts and
measurements in descriptive and analytic studies, stratified analysis with
exact confidence limits, matched pair and person-time analysis, sample size
and power calculations, random numbers, sensitivity, specificity and other
evaluation statistics, R x C tables and chi-square for dose-response. The
original software was written in JavaScript and HTML and was designed to be
a serverless, browser-based application that can be used on any computer
platform including smartphones. With the advent of Shiny in R, re-creating
the OpenEpi browser-based application has become possible. This package
translates the JavaScript calculator functions used in the original
application into R and includes a Shiny application that uses these functions
to replicate the original OpenEpi browser application.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [contributor]

Authors:
\itemize{
  \item Dean Andrew [copyright holder]
  \item Sullivan Kevin [copyright holder]
  \item Soe Minn [copyright holder]
}

}
\keyword{internal}
